% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMPSS-package.R
\docType{package}
\name{RMPSS-package}
\alias{RMPSS}
\alias{RMPSS-package}
\title{RMPSS: Recursive Modified Pattern Search on Simplex}
\description{
Pattern search based Blackbox optimization algorithm designed to optimize any function with parameters belonging to unit simplex. Described in the paper : Das (2016) <arXiv:1604.08636> .
}
\author{
\strong{Maintainer}: Priyam Das \email{pdas@ncsu.edu}

Authors:
\itemize{
  \item Debsurya De \email{debsurya001@gmail.com}
}

}
\keyword{internal}
